+++ R/CohortConstruction.R
@@ -16,7 +16,7 @@
 
 getCohortsJsonAndSqlFromPackage <- function(packageName = packageName,
                                             cohortToCreateFile = cohortToCreateFile,
-                                            cohortIds = cohortIds,
+                                            cohortIds = NULL,
                                             errorMessage = NULL) {
   ParallelLogger::logInfo("Executing on cohorts specified in package - ", packageName)
   
@@ -30,8 +30,10 @@ getCohortsJsonAndSqlFromPackage <- function(packageName = packageName,
                               extension = "csv", 
                               add = errorMessage)
   
-  cohorts <- readr::read_csv(pathToCsv, col_types = readr::cols()) %>%
-    dplyr::filter(!.data$cohortId %in% cohortIds)
+  cohorts <- readr::read_csv(pathToCsv, col_types = readr::cols())
+  if (!is.null(cohortIds)) {
+    cohorts <- cohorts %>% dplyr::filter(.data$cohortId %in% cohortIds)
+  }
   
   checkmate::assertDataFrame(x = cohorts, 
                              types = c("integer", "character","numeric"),
@@ -66,10 +68,9 @@ getCohortsJsonAndSqlFromPackage <- function(packageName = packageName,
 }
 
 
-
 getCohortsJsonAndSqlFromWebApi <- function(baseUrl = baseUrl,
                                            cohortSetReference = cohortSetReference,
-                                           cohortIds = cohortIds,
+                                           cohortIds = NULL,
                                            errorMessage = NULL) {
   ParallelLogger::logInfo("[WebApi mode] Running Cohort Diagnostics on cohort specified in WebApi - ", baseUrl)
   
@@ -92,9 +93,11 @@ getCohortsJsonAndSqlFromWebApi <- function(baseUrl = baseUrl,
                          subset.of =  c("atlasName", "atlasId", "cohortId", "name", "cohortName"),
                          add = errorMessage)
   checkmate::reportAssertions(collection = errorMessage)
-  cohorts <- cohortSetReference %>%
-    dplyr::filter(!.data$cohortId %in% cohortIds)
-  
+  cohorts <- cohortSetReference
+  if (!is.null(cohortIds)) {
+    cohorts <- cohorts %>% dplyr::filter(.data$cohortId %in% cohortIds)
+  }
+
   if ("name" %in% names(cohorts)) {
     cohorts <- dplyr::rename(cohorts, cohortName = "name")
   }
@@ -766,6 +769,7 @@ saveAndDropTempInclusionStatsTables <- function(connection,
                                                        oracleTempSchema = oracleTempSchema,
                                                        snakeCaseToCamelCase = TRUE,
                                                        table = table)
+    data <- tidyr::as_tibble(data)
     fullFileName <- file.path(inclusionStatisticsFolder, fileName)
     if (incremental) {
       saveIncremental(data, fullFileName, cohortDefinitionId = cohortIds)